requirements.txt
--index-url https://download.pytorch.org/whl/cu117
torch==2.0.0+cu117
torchvision==0.15.1+cu117
Pillow>=9.0.0
numpy>=1.17.5
tqdm
matplotlib
opencv-python